Split test_skill.py into thematic modules for readability
Move data models/config to config.py, generic helpers to utils.py,
and test environment setup/teardown to infra.py. test_skill.py now
contains only the core test logic (~270 lines, down from ~860).
